Rename game dir to avoid bug in playsound module. Fix requirements.txt to include valid modules.
