--- Game Files/requirements.txt
@@ -1,2 +1,0 @@
-pkg-resources==0.0.0
-playsound==1.2.2
