+++ game/requirements.txt
@@ -1,0 +1,1 @@
+playsound==1.2.2
